requirements.txt
fal-client
requests
pathlib
PyQt6
python-dotenv
google-generativeai